refactor: extract home line summary skeleton
